requirements-dev.txt
ansible-core>=2.12
ansible-lint
black
click
filelock
flake8
iac-validate>=0.1.3
jmespath
jsonpath_ng
mkdocs
mkdocs-material
mkdocs-mermaid2-plugin
molecule
pyopenssl
pytest
pytest-mock
python-dateutil
PyVmomi
requests
requests-toolbelt
ruamel.yaml
robotframework>=4.0
robotframework-pabot
robotframework-jsonlibrary>=0.5
robotframework-requests>=0.9.3
yamllint
